Изменен код в модуле main
