+++ src/main.py
@@ -1,105 +1,87 @@
+from src.category import Category
 from src.smartphone import Smartphone
 from src.lawngrass import LawnGrass
 
-class Product:
-    """Класс продуктов"""
-
-    name: str
-    description: str
-    price: float
-    quantity: int
-
-    def __init__(self, name, description, price, quantity):
-        self.name = name
-        self.description = description
-        self.__price = price  # Приватный атрибут для цены
-        self.quantity = quantity
-
-    @property
-    def price(self):
-        """Геттер для цены"""
-        return self.__price
-
-    @price.setter
-    def price(self, value):
-        """Сеттер для цены"""
-        if value <= 0:
-            print("Цена не должна быть нулевая или отрицательная")
-        else:
-            self.__price = value
-
-    @classmethod
-    def new_product(cls, product_info):
-        """Создает новый объект Product на основе словаря с параметрами"""
-        name = product_info.get("name")
-        description = product_info.get("description")
-        price = product_info.get("price")
-        quantity = product_info.get("quantity")
-        return cls(name, description, price, quantity)
-
-    def __str__(self):
-        return f"{self.name}, {self.price} руб. Остаток: {self.quantity} шт."
-
-    def __add__(self, other):
-        if isinstance(other, Product):
-            if type(self) == type(other):
-                return (self.price * self.quantity) + (other.price * other.quantity)
-            else:
-                TypeError("Нельзя складывать продукты разных классов")
-        return NotImplemented
 
 
-class Category:
-    """Класс категории продуктов"""
+if __name__ == "__main__":
 
-    name: str
-    description: str
-    products: list[dict]
+    smartphone1 = Smartphone("Samsung Galaxy S23 Ultra", "256GB, Серый цвет, 200MP камера", 180000.0, 5, 95.5,
+                             "S23 Ultra", 256, "Серый")
+    smartphone2 = Smartphone("Iphone 15", "512GB, Gray space", 210000.0, 8, 98.2, "15", 512, "Gray space")
+    smartphone3 = Smartphone("Xiaomi Redmi Note 11", "1024GB, Синий", 31000.0, 14, 90.3, "Note 11", 1024, "Синий")
 
-    category_count = 0
-    product_count = 0
+    print(smartphone1.name)
+    print(smartphone1.description)
+    print(smartphone1.price)
+    print(smartphone1.quantity)
+    print(smartphone1.efficiency)
+    print(smartphone1.model)
+    print(smartphone1.memory)
+    print(smartphone1.color)
 
-    def __init__(self, name, description, products):
-        self.name = name
-        self.description = description
-        self.__products = products  # Приватный атрибут для хранения продуктов
-        Category.category_count += 1  # Увеличиваем счетчик категорий при создании нового объекта
+    print(smartphone2.name)
+    print(smartphone2.description)
+    print(smartphone2.price)
+    print(smartphone2.quantity)
+    print(smartphone2.efficiency)
+    print(smartphone2.model)
+    print(smartphone2.memory)
+    print(smartphone2.color)
 
-    def add_product(self, product):
-        """Добавляет продукт в категорию"""
-        if isinstance(product, Product):
-            self.__products.append(product)
-            Category.product_count += product.quantity  # Увеличиваем общий счетчик на количество добавленного продукта
+    print(smartphone3.name)
+    print(smartphone3.description)
+    print(smartphone3.price)
+    print(smartphone3.quantity)
+    print(smartphone3.efficiency)
+    print(smartphone3.model)
+    print(smartphone3.memory)
+    print(smartphone3.color)
 
-    @property
-    def products(self):
-        """Возвращает список продуктов в виде строки"""
-        return "\n".join(str(product) for product in self.__products)
+    grass1 = LawnGrass("Газонная трава", "Элитная трава для газона", 500.0, 20, "Россия", "7 дней", "Зеленый")
+    grass2 = LawnGrass("Газонная трава 2", "Выносливая трава", 450.0, 15, "США", "5 дней", "Темно-зеленый")
 
-    def __str__(self):
-        total_quantity = sum(product.quantity for product in self.__products)
-        return f"{self.name}, количество продуктов: {total_quantity} шт."
+    print(grass1.name)
+    print(grass1.description)
+    print(grass1.price)
+    print(grass1.quantity)
+    print(grass1.country)
+    print(grass1.germination_period)
+    print(grass1.color)
 
+    print(grass2.name)
+    print(grass2.description)
+    print(grass2.price)
+    print(grass2.quantity)
+    print(grass2.country)
+    print(grass2.germination_period)
+    print(grass2.color)
 
-# if __name__ == "__main__":
-#     product1 = Product("Samsung Galaxy S23 Ultra", "256GB, Серый цвет, 200MP камера", 180000.0, 5)
-#     product2 = Product("Iphone 15", "512GB, Gray space", 210000.0, 8)
-#     product3 = Product("Xiaomi Redmi Note 11", "1024GB, Синий", 31000.0, 14)
-#
-#     print(str(product1))
-#     print(str(product2))
-#     print(str(product3))
-#
-#     category1 = Category(
-#         "Смартфоны",
-#         "Смартфоны, как средство не только коммуникации, но и получения дополнительных функций для удобства жизни",
-#         [product1, product2, product3]
-#     )
-#
-#     print(str(category1))
-#
-#     print(category1.products)
-#
-#     print(product1 + product2)
-#     print(product1 + product3)
-#     print(product2 + product3)
+    smartphone_sum = smartphone1 + smartphone2
+    print(smartphone_sum)
+
+    grass_sum = grass1 + grass2
+    print(grass_sum)
+
+    try:
+        invalid_sum = smartphone1 + grass1
+    except TypeError:
+        print("Возникла ошибка TypeError при попытке сложения")
+    else:
+        print("Не возникла ошибка TypeError при попытке сложения")
+
+    category_smartphones = Category("Смартфоны", "Высокотехнологичные смартфоны", [smartphone1, smartphone2])
+    category_grass = Category("Газонная трава", "Различные виды газонной травы", [grass1, grass2])
+
+    category_smartphones.add_product(smartphone3)
+
+    print(category_smartphones.products)
+
+    print(Category.product_count)
+
+    try:
+        category_smartphones.add_product("Not a product")
+    except TypeError:
+        print("Возникла ошибка TypeError при добавлении не продукта")
+    else:
+        print("Не возникла ошибка TypeError при добавлении не продукта")
